python/pyproject.toml
[project]
name = "smart-chatbot"
version = "1.0.0"
description = "Hybrid AI Chatbot with n8n + Python for Facebook Messenger & Zalo"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Smart Chatbot Team"}
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    "redis>=5.0.0",
    "neo4j>=5.15.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langchain-community>=0.0.10",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "tenacity>=8.2.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "httpx>=0.26.0",
]
production = [
    "slowapi>=0.1.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
